code refactoring and cleanup
