+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/opmodes/autonomous/Test.java
@@ -1,4 +1,4 @@
-package org.firstinspires.ftc.teamcode.opmodes;
+package org.firstinspires.ftc.teamcode.opmodes.autonomous;
 
 
 import android.util.Log;
