+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/opmodes/autonomous/TestDrive.java
@@ -1,19 +1,16 @@
-package org.firstinspires.ftc.teamcode.opmodes;
+package org.firstinspires.ftc.teamcode.opmodes.autonomous;
 
-
-import android.util.Log;
 
 import com.qualcomm.robotcore.eventloop.opmode.Autonomous;
 import com.qualcomm.robotcore.util.ElapsedTime;
 
-import org.firstinspires.ftc.robotcore.external.matrices.VectorF;
-import org.firstinspires.ftc.robotcore.external.navigation.Orientation;
 import org.firstinspires.ftc.robotcore.external.navigation.VuforiaTrackable;
 import org.firstinspires.ftc.teamcode.components.DriveSystem;
 import org.firstinspires.ftc.teamcode.components.Vuforia;
+import org.firstinspires.ftc.teamcode.opmodes.base.BaseOpMode;
 
 @Autonomous(name = "TestDrive", group="Autonomous")
-public class TestDrive extends BaseStateMachine {
+public class TestDrive extends BaseOpMode {
     public enum State {
         STATE_INITIAL,
         STATE_RIGHT,
@@ -23,14 +20,9 @@ public class TestDrive extends BaseStateMachine {
         STATE_FINISHED,
     }
 
-    public ElapsedTime  mRuntime = new ElapsedTime();   // Time into round.
-
     public static final String TAG = "Test Drive";
 
-    private ElapsedTime mStateTime = new ElapsedTime();  // Time into current state
-    private Vuforia vuforia;
-    private VuforiaTrackable skystone;
-    private DriveSystem driveSystem;
+
     private static final float mmPerInch = 25.4f;
 
     protected State mCurrentState;    // Current State Machine State.
@@ -38,7 +30,6 @@ public class TestDrive extends BaseStateMachine {
     @Override
     public void init() {
         super.init();
-        driveSystem = super.driveSystem;
         mCurrentState = State.STATE_INITIAL;
     }
 
@@ -52,22 +43,22 @@ public class TestDrive extends BaseStateMachine {
                 break;
 
             case STATE_FORWARD:
-                if(driveSystem.driveToPositionInches(48, DriveSystem.Direction.FORWARD, .5)){
+                if(driveSystem.driveToPosition(100, DriveSystem.Direction.FORWARD, .5)){
                     newState(State.STATE_BACKWARD);
                 }
                 break;
             case STATE_BACKWARD:
-                if(driveSystem.driveToPositionInches(48, DriveSystem.Direction.BACKWARD, .5)){
+                if(driveSystem.driveToPosition(100, DriveSystem.Direction.BACKWARD, .5)){
                     newState(State.STATE_RIGHT);
                 }
                 break;
             case STATE_RIGHT:
-                if(driveSystem.driveToPositionInches(48, DriveSystem.Direction.RIGHT, .5)){
+                if(driveSystem.driveToPosition(100, DriveSystem.Direction.RIGHT, .5)){
                     newState(State.STATE_LEFT);
                 }
                 break;
             case STATE_LEFT:
-                if(driveSystem.driveToPositionInches(48, DriveSystem.Direction.LEFT, .5)){
+                if(driveSystem.driveToPosition(100, DriveSystem.Direction.LEFT, .5)){
                     newState(State.STATE_FINISHED);
                 }
                 break;
@@ -82,8 +73,6 @@ public class TestDrive extends BaseStateMachine {
     }
 
     public void newState(State newState) {
-        // Restarts the state clock as well as the state
-        mStateTime.reset();
         mCurrentState = newState;
     }
